+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Helm: support labels in values.yaml ([#344]).
 - Support version `3.4.2` ([#357]).
+- `spec.job.config.volumeMounts` property to easily mount volumes on the job pod ([#359])
 
 ### Changed
 
@@ -29,6 +30,7 @@ All notable changes to this project will be documented in this file.
 [#344]: https://github.com/stackabletech/spark-k8s-operator/pull/344
 [#355]: https://github.com/stackabletech/spark-k8s-operator/pull/355
 [#357]: https://github.com/stackabletech/spark-k8s-operator/pull/357
+[#359]: https://github.com/stackabletech/spark-k8s-operator/pull/359
 
 ## [23.11.0] - 2023-11-24